Extend the schema compatibility example to Image
The drift is not Work-specific: Image carries requestUrl, which the API
does not return (it is a client-side augmentation from the fetch layer's
_requestUrl), and its inline source omits fields the schema documents.
